no teardown functionality

+++ tests/global/teardown.js
@@ -7,9 +7,14 @@ const TEARDOWN_OUT_DIRECTORY = path.join(__dirname, "../out/");
 module.exports = function (globalConfig, projectConfig) {
   let dirs = [TEARDOWN_DATA_FILEPATH, TEARDOWN_OUT_DIRECTORY];
 
-  dirs.forEach((dir) => {
-    fs.rmSync(dir, { recursive: true, force: true });
-  });
+  if (projectConfig.globals.TEARDOWN) {
+    dirs.forEach((dir) => {
+      fs.rmSync(dir, { recursive: true, force: true });
+    });
 
-  console.log("### Teardown complete ###");
+    console.log(`### Teardown complete ###`);
+  } else {
+    console.log("### Teardown did not occur ###");
+    console.log("change jest global variable TEARDOWN to true (package.json)");
+  }
 };
